# Refactored deepMerge function to improve performance