--- a/.changeset/late-coins-beam.md
+++ b/.changeset/late-coins-beam.md
@@ -1,0 +1,5 @@
+---
+'svelte-meta-tags': patch
+---
+
+Refactored deepMerge function to improve performance