--- a/.github/workflows/ci.yml
+++ b/.github/workflows/ci.yml
@@ -24,5 +24,5 @@ jobs:
         run: pnpm build
       - name: Install Playwright
         run: pnpm dlx playwright@1.48.1 install --with-deps
-      - name: Run Playwright
+      - name: Run Vitest and Playwright
         run: pnpm test